+++ core/src/main/kotlin/org/evomaster/core/problem/enterprise/auth/AuthSettings.kt
@@ -1,5 +1,7 @@
 package org.evomaster.core.problem.enterprise.auth
 
+import org.evomaster.core.search.service.Randomness
+
 
 /**
  * Class used to store and handle all Auth Info for the current SUT
@@ -28,4 +30,18 @@ class AuthSettings(authInfos: List<AuthenticationInfo> = listOf()) {
     fun isNotEmpty() = auths.isNotEmpty()
 
     fun size() = auths.size
+
+    fun <T : AuthenticationInfo> getDifferentOne(name: String, klass: Class<T>, randomness: Randomness) : T {
+
+        if(auths.none { it.name == name }){
+            throw IllegalArgumentException("No auth info with name: $name")
+        }
+
+        val others = getOfType(klass).filter { it.name != name }
+        if(others.isEmpty()){
+            throw IllegalArgumentException("There is no other auth config besides '$name' for type $klass")
+        }
+
+        return randomness.choose(others)
+    }
 }
