+++ core/src/main/kotlin/org/evomaster/core/problem/enterprise/auth/NoAuth.kt
@@ -1,0 +1,4 @@
+package org.evomaster.core.problem.enterprise.auth
+
+
+interface NoAuth
