+++ core/src/main/kotlin/org/evomaster/core/problem/httpws/auth/HttpWsNoAuth.kt
@@ -1,0 +1,6 @@
+package org.evomaster.core.problem.httpws.auth
+
+import org.evomaster.core.problem.enterprise.auth.NoAuth
+
+
+class HttpWsNoAuth : HttpWsAuthenticationInfo("NoAuth", listOf(), null, null), NoAuth
