--- core/src/main/kotlin/org/evomaster/core/problem/httpws/auth/NoAuth.kt
@@ -1,4 +1,0 @@
-package org.evomaster.core.problem.httpws.auth
-
-
-class NoAuth : HttpWsAuthenticationInfo("NoAuth", listOf(), null, null)
